+++ tests/codeigniter/helpers/string_helper_test.php
@@ -98,7 +98,7 @@ class String_helper_test extends CI_TestCase {
 	public function test_random_string()
 	{
 		$this->assertEquals(16, strlen(random_string('alnum', 16)));
-		$this->assertEquals(32, strlen(random_string('unique', 16)));
+		$this->assertEquals(32, strlen(random_string('md5', 16)));
 		$this->assertEquals('string', gettype(random_string('numeric', 16)));
 	}