+++ tests/codeigniter/helpers/url_helper_test.php
@@ -21,7 +21,7 @@ class Url_helper_test extends CI_TestCase {
 
 		foreach ($words as $in => $out)
 		{
-			$this->assertEquals($out, url_title($in, 'dash', TRUE));
+			$this->assertEquals($out, url_title($in, '-', TRUE));
 		}
 	}
 
@@ -41,7 +41,7 @@ class Url_helper_test extends CI_TestCase {
 
 		foreach ($words as $in => $out)
 		{
-			$this->assertEquals($out, url_title($in, 'underscore'));
+			$this->assertEquals($out, url_title($in, '_'));
 		}
 	}